project controller

+++ src/controllers/project.controllers.js
@@ -4,7 +4,7 @@ import ApiResponse from '../utils/api-response.js';
 import { ApiError } from '../utils/api-errors.js';
 import { User } from '../models/user.models.js';
 import { ProjectMember } from '../models/projectmember.models.js';
-import {UserRolesEnum} from '../utils/contants.js'
+import {UserRolesEnum, AvailableUserRoles, ProjectStatusEnum} from '../utils/contants.js'
 
 const getProjects = asyncHandler(async (req, res) => {
   console.log('=====getProjects controller=====');
@@ -36,19 +36,16 @@ const getProjectById = asyncHandler(async (req, res) => {
     throw new ApiError(401, 'Unauthorized request! Please login first.');
   }
   try {
-    const project = await Project.findById({ _id: projectId });
+    const project = await Project.findById({ _id: projectId }).populate("createdBy", 'fname username email');
     if (!project) {
       throw new ApiError(404, 'Project not found!');
     }
 
     return res.status(200).json(
       new ApiResponse(
-        200,
-        {
-          project: project.name,
-          description: project.description,
-          createdBy: project.createdBy,
-        },
+        200,        
+        project
+        ,
         'Project found successfully!',
       ),
     );
@@ -57,6 +54,7 @@ const getProjectById = asyncHandler(async (req, res) => {
   }
 });
 const createProject = asyncHandler(async (req, res) => {
+  // await Project.deleteMany({});
   try {
     const { name, description } = req.body;
     const userId = req.user._id;
@@ -76,34 +74,37 @@ const createProject = asyncHandler(async (req, res) => {
       name,
       description,
       createdBy: userId,
-    }.populate("createdBy"));
+      status: ProjectStatusEnum.in_progress
+    });
 
     if (!project) {
+      throw new ApiError(400, 'Project could not be created!');
+    }
+    const projectPopulated = await Project.findById(project._id).populate('createdBy', 'fname username email');
+    if(!projectPopulated){
       throw new ApiError(400, 'Project could not be created!');
     }
     const projectMember = await ProjectMember.create({
       user: userId,
       project: project._id,  
-      role: UserRolesEnum.ADMIN
+      role: UserRolesEnum.PROJECT_ADMIN
     });
 
     if (!projectMember) {
       throw new ApiError(400, 'Project member could not be created!');
     }
-    projectMember.save();
-    project.save();
-
-    return res.status(200).json(new ApiResponse(200, project, 'Project created successfully!'));
+    
+    return res.status(200).json(new ApiResponse(200, projectPopulated, 'Project created successfully!'));
   } catch (error) {
     res.status(400).json(new ApiResponse(400, error, error.message));
   }
 });
 const updateProject = asyncHandler(async (req, res) => {
   const userId = req.user._id;
-  const { id } = req.query;
-  const { newName, newDescription } = req.body;
+  const { projectId: id } = req.params;
+  const { name, description, status } = req.body;  
 
-  if (!newName || !newDescription) {
+  if (!name || !description) {
     throw new ApiError(400, 'Please fill all the fields');
   }
   if (!id) {
@@ -123,15 +124,21 @@ const updateProject = asyncHandler(async (req, res) => {
     throw new ApiError(403, 'You are not authorized to update this project!');
   }
 
-  project.name = newName;
-  project.description = newDescription;
+  const updatedProject = await Project.findByIdAndUpdate(
+    id,
+    { name: name, description: description, status: status },
+    { new: true },
+  );
+  if (!updatedProject) {
+    throw new ApiError(400, 'Project could not be updated!');
+  }
   await project.save();
 
-  return res.status(200).json(new ApiResponse(200, project, 'Project found successfully!'));
+  return res.status(200).json(new ApiResponse(200, updatedProject, 'Project updated successfully!'));
 });
 const deleteProject = asyncHandler(async (req, res) => {
   const userId = req.user._id;
-  const { id } = req.query;
+  const { projectId: id } = req.params;
 
   if (!userId) {
     throw new ApiError(401, 'Unauthorized request! Please login first.');
@@ -140,25 +147,31 @@ const deleteProject = asyncHandler(async (req, res) => {
     throw new ApiError(400, 'Please provide a project id!');
   }
 
-  // new
-  const projectMember = await ProjectMember.findOne({ user: userId, project: id });
-  if (!projectMember) {
-    throw new ApiError(403, 'You are not a member of this project!');
+  const projectMember = await ProjectMember.findOne({ user: userId, project: id }); 
+  
+  if (!projectMember || !projectMember.role) {
+    throw new ApiError(403, 'You are not a member of this project or do not have the required role!');
   }
-  if (projectMember.role !== 'member') {
+  
+  if (projectMember.role !== 'project_admin') {
     throw new ApiError(403, 'You are not authorized to delete this project!');
   }
-  const projectDelete = await Project.findByIdAndDelete(id);
-  if (!projectDelete) {
-    throw new ApiError(404, 'Project could not be deleted!');
-  }
-  const deletedMembers = await ProjectMember.deleteMany({ project: id });
 
-  if (!deletedMembers) {
-    throw new ApiError(404, 'Project members could not be deleted!');
-  }
+  try {
+    const projectDelete = await Project.findByIdAndDelete(id);
+    if (!projectDelete) {
+      throw new ApiError(404, 'Project could not be deleted!');
+    }
 
-  return res.status(200).json(new ApiResponse(200, projectDelete, 'Project deleted successfully!'));
+    const deletedMembers = await ProjectMember.deleteMany({ project: id }, { new: true });
+    if (!deletedMembers || deletedMembers.deletedCount === 0) {
+      throw new ApiError(404, 'Project members could not be deleted!');
+    }
+
+    return res.status(200).json(new ApiResponse(200, projectDelete, 'Project deleted successfully!'));
+  } catch (error) {
+    return res.status(500).json(new ApiError(500,'An error occurred while deleting the project!'));
+  }
 });
 const addMemberToProject = asyncHandler(async (req, res) => {
   const userId = req.user._id;
@@ -176,7 +189,7 @@ const addMemberToProject = asyncHandler(async (req, res) => {
   if (!project) {
     throw new ApiError(404, 'Project not found!');
   }
-  if (userId.toString() !== project.createdBy.toString()) {
+  if (!project.createdBy.equals(userId)) {
     throw new ApiError(403, 'You are not authorized to add members to this project!');
   }
   // email checking
@@ -200,17 +213,16 @@ const addMemberToProject = asyncHandler(async (req, res) => {
   const projectMember = await ProjectMember.create({
     user: user._id,
     project: projectId,
-    role: 'member',
+    role: UserRolesEnum.MEMBER,
   });
 
   if (!projectMember) {
     throw new ApiError(400, 'Project member could not be created!');
   }
-  await projectMember.save();
-
+ 
   return res
     .status(200)
-    .json(new ApiResponse(200, { projectMember }, 'Member added to project successfully!'));
+    .json(new ApiResponse(200,  projectMember , 'Member added to project successfully!'));
 });
 const getProjectMembers = asyncHandler(async (req, res) => {
   const userId = req.user._id;
@@ -229,7 +241,7 @@ const getProjectMembers = asyncHandler(async (req, res) => {
   if (userId.toString() !== project.createdBy.toString()) {
     throw new ApiError(403, 'You are not authorized to get members of this project!');
   }
-  const projectMembers = await ProjectMember.find({ project: projectId });
+  const projectMembers = await ProjectMember.find({ project: projectId }).populate('user');
   if (!projectMembers) {
     throw new ApiError(404, 'Project members could not be found!');
   }
@@ -285,8 +297,11 @@ const updateProjectMembers = asyncHandler(async (req, res) => {
     .json(new ApiResponse(200, updateProject, 'Project members updated successfully!'));
 });
 const updateMemberRole = asyncHandler(async (req, res) => {
+  console.log("=========updateMemberRole controller==========");
   const userId = req.user._id;
   const { projectId, memberId } = req.params;
+  const { role } = req.body;  
+  
   if (!userId) {
     throw new ApiError(401, 'Unauthorized request! Please login first.');
   }
@@ -301,7 +316,7 @@ const updateMemberRole = asyncHandler(async (req, res) => {
   if (!project) {
     throw new ApiError(404, 'Project not found!');
   }
-  if (userId.toString() !== project.createdBy.toString()) {
+  if (!project.createdBy.equals(userId)) {
     throw new ApiError(403, 'You are not authorized to update members of this project!');
   }
 
@@ -323,10 +338,10 @@ const updateMemberRole = asyncHandler(async (req, res) => {
     throw new ApiError(403, 'You are not authorized to update members of this project!');
   }
 
-  if (!req.body.role) {
+  if (!role) {
     throw new ApiError(400, 'Please provide a role!');
   }
-  projectMember.role = req.body.role;
+  projectMember.role = role;
   await projectMember.save();
 
   return res
@@ -335,6 +350,7 @@ const updateMemberRole = asyncHandler(async (req, res) => {
 });
 
 const deleteMember = asyncHandler(async (req, res) => {
+  console.log("=========deleteMember controller==========");
   const userId = req.user._id;
   const { projectId, memberId } = req.params;
   if (!userId) {
@@ -351,7 +367,7 @@ const deleteMember = asyncHandler(async (req, res) => {
   if (!project) {
     throw new ApiError(404, 'Project not found!');
   }
-  if (userId.toString() !== project.createdBy.toString()) {
+  if (!project.createdBy.equals(userId)) {
     throw new ApiError(403, 'You are not authorized to update members of this project!');
   }
   
@@ -361,23 +377,23 @@ const deleteMember = asyncHandler(async (req, res) => {
   });
   if (!superAdmin) {
     throw new ApiError(404, 'Project members could not be found!');
-  }
-  superAdmin.role = 'admin';
-  if (superAdmin.role !== 'admin') {
+  }  
+  if (superAdmin.role !== 'project_admin') {
     throw new ApiError(403, 'You are not authorized to update members of this project!');
   }
   const removeMember = await ProjectMember.deleteOne({
-    project: projectId,
     user: memberId,
+    project: projectId,
   }, { new: true });
-
+  console.log(removeMember);
+  
   if (!removeMember) {
     throw new ApiError(404, 'Project members could not be found!');
   }
 
   return res
     .status(200)
-    .json(new ApiResponse(200, removeMember, 'Project members updated successfully!'));
+    .json(new ApiResponse(200, removeMember, 'project members deleted successfully!'));
 });
 
 export {
